+++ src/versions/v2/models/get-deals-products-response-data-inner-all-of1.ts
@@ -21,7 +21,7 @@
 */
 export interface GetDealsProductsResponseDataInnerAllOf1 {
     /**
-    * Only available in Advanced and above plans  How often a customer is billed for access to a service or product  To set `billing_frequency` different than `one-time`, the deal must not have installments associated  A deal can have up to 20 products attached with `billing_frequency` different than `one-time` 
+    * Only available in Growth and above plans  How often a customer is billed for access to a service or product  To set `billing_frequency` different than `one-time`, the deal must not have installments associated  A deal can have up to 20 products attached with `billing_frequency` different than `one-time` 
     * @type {string}
     */
     'billing_frequency'?: GetDealsProductsResponseDataInnerAllOf1BillingFrequencyConst;